all the output in get_para_report.csv
